+++ dojo/api_v2/serializers.py
@@ -2,9 +2,10 @@ from typing import List
 from drf_spectacular.utils import extend_schema_field
 from drf_yasg.utils import swagger_serializer_method
 from rest_framework.fields import DictField, MultipleChoiceField
+from dojo.api_v2.permissions import get_import_meta_data_from_dict
 
 from dojo.endpoint.utils import endpoint_filter
-from dojo.importers.reimporter.utils import get_import_meta_data_from_dict, get_target_engagement_if_exists, get_target_product_if_exists, get_target_test_if_exists
+from dojo.importers.reimporter.utils import get_target_engagement_if_exists, get_target_product_if_exists, get_target_test_if_exists
 from dojo.models import Dojo_User, Finding_Group, Product, Engagement, Test, Finding, \
     User, Stub_Finding, Risk_Acceptance, \
     Finding_Template, Test_Type, Development_Environment, NoteHistory, \
@@ -1230,9 +1231,9 @@ class ImportScanSerializer(serializers.Serializer):
 
         group_by = data.get('group_by', None)
 
-        _, test_title, scan_type, engagement_id, engagement_name, product_id, product_name = get_import_meta_data_from_dict(data)
+        _, test_title, scan_type, engagement_id, engagement_name, product_name = get_import_meta_data_from_dict(data)
         # we passed validation, so the engagement is present
-        product = get_target_product_if_exists(product_id, product_name)
+        product = get_target_product_if_exists(product_name)
         engagement = get_target_engagement_if_exists(engagement_id, engagement_name, product)
 
         importer = Importer()
@@ -1259,7 +1260,6 @@ class ImportScanSerializer(serializers.Serializer):
         # return the id of the created test, can't find a better way because this is not a ModelSerializer....
         if test:
             self.fields['test'] = serializers.IntegerField(read_only=True, default=test.id)
-            self.fields['id'] = serializers.IntegerField(read_only=True, default=test.id)
             data['engagement'] = test.engagement
             data['product'] = test.engagement.product.id
 
@@ -1344,9 +1344,9 @@ class ReImportScanSerializer(TaggitSerializer, serializers.Serializer):
 
         group_by = data.get('group_by', None)
 
-        test_id, test_title, scan_type, _, engagement_name, product_id, product_name = get_import_meta_data_from_dict(data)
+        test_id, test_title, scan_type, _, engagement_name, product_name = get_import_meta_data_from_dict(data)
         # we passed validation, so the engagement is present
-        product = get_target_product_if_exists(product_id, product_name)
+        product = get_target_product_if_exists(product_name)
         engagement = get_target_engagement_if_exists(None, engagement_name, product)
         test = get_target_test_if_exists(test_id, test_title, scan_type, engagement)
 
@@ -1370,7 +1370,6 @@ class ReImportScanSerializer(TaggitSerializer, serializers.Serializer):
         # return the id of the created test, can't find a better way because this is not a ModelSerializer....
         if test:
             data['test'] = test
-            data['id'] = test.id
             data['engagement'] = test.engagement
             data['product'] = test.engagement.product.id
